+++ asdf/generic_io.py
@@ -213,12 +213,17 @@ class GenericFile(metaclass=util.InheritDocstrings):
             raise ValueError(
                 "File-like object must be opened in binary mode.")
 
+        # can't import at the top level due to circular import
+        from .config import get_config
+        self._asdf_get_config = get_config
+
         self._fd = fd
         self._mode = mode
         self._close = close
-        self._blksize = io.DEFAULT_BUFFER_SIZE
         self._size = None
         self._uri = uri
+
+        self.block_size = get_config().io_block_size
 
     def __enter__(self):
         return self
@@ -233,6 +238,19 @@ class GenericFile(metaclass=util.InheritDocstrings):
     @property
     def block_size(self):
         return self._blksize
+
+    @block_size.setter
+    def block_size(self, block_size):
+        if block_size == -1:
+            try:
+                block_size = os.fstat(self._fd.fileno()).st_blksize
+            except Exception:
+                block_size = io.DEFAULT_BUFFER_SIZE
+
+        if block_size <= 0:
+            raise ValueError(f'block_size ({block_size}) must be > 0')
+
+        self._blksize = block_size
 
     @property
     def mode(self):
@@ -668,13 +686,8 @@ class RealFile(RandomAccessFile):
     """
     def __init__(self, fd, mode, close=False, uri=None):
         super(RealFile, self).__init__(fd, mode, close=close, uri=uri)
+
         stat = os.fstat(fd.fileno())
-        if sys.platform.startswith('win'):  # pragma: no cover
-            # There appears to be reliable way to get block size on Windows,
-            # so just choose a reasonable default
-            self._blksize = io.DEFAULT_BUFFER_SIZE
-        else:
-            self._blksize = stat.st_blksize
         self._size = stat.st_size
         if (uri is None and
             isinstance(fd.name, str)):
@@ -828,7 +841,7 @@ class HTTPConnection(RandomAccessFile):
 
     def __init__(self, connection, size, path, uri, first_chunk):
         self._mode = 'r'
-        self._blksize = io.DEFAULT_BUFFER_SIZE
+        self.block_size = io.DEFAULT_BUFFER_SIZE
         # The underlying HTTPConnection object doesn't track closed
         # status, so we do that here.
         self._closed = False