+++ asdf/tests/test_extension.py
@@ -10,6 +10,7 @@ from asdf.extension import (
     TagDefinition,
     Converter,
     ConverterProxy,
+    Compressor,
     AsdfExtension,
     BuiltinExtension,
     get_cached_asdf_extension_list
@@ -52,11 +53,13 @@ class FullExtension:
     def __init__(
         self,
         converters=None,
+        compressors=None,
         asdf_standard_requirement=None,
         tags=None,
         legacy_class_names=None,
     ):
         self._converters = [] if converters is None else converters
+        self._compressors = [] if compressors is None else compressors
         self._asdf_standard_requirement = asdf_standard_requirement
         self._tags = tags
         self._legacy_class_names = [] if legacy_class_names is None else legacy_class_names
@@ -64,6 +67,10 @@ class FullExtension:
     @property
     def converters(self):
         return self._converters
+
+    @property
+    def compressors(self):
+        return self._compressors
 
     @property
     def asdf_standard_requirement(self):
@@ -110,6 +117,14 @@ class FullConverter(MinimumConverter):
         return "select_tag result"
 
 
+class MinimalCompressor(Compressor):
+    def compress(data):
+        return b''
+
+    @property
+    def label(self):
+        return b'mini'
+
 # Some dummy types for testing converters:
 class FooType:
     pass
@@ -145,6 +160,7 @@ def test_extension_proxy():
     assert proxy.legacy_class_names == set()
     assert proxy.asdf_standard_requirement == SpecifierSet()
     assert proxy.converters == []
+    assert proxy.compressors == []
     assert proxy.tags == []
     assert proxy.types == []
     assert proxy.tag_mapping == []
@@ -181,6 +197,7 @@ def test_extension_proxy():
     ]
     extension = FullExtension(
         converters=converters,
+        compressors=[MinimalCompressor],
         asdf_standard_requirement=">=1.4.0",
         tags=["asdf://somewhere.org/extensions/full/tags/foo-1.0"],
         legacy_class_names=["foo.extensions.SomeOldExtensionClass"]
@@ -191,6 +208,7 @@ def test_extension_proxy():
     assert proxy.legacy_class_names == {"foo.extensions.SomeOldExtensionClass"}
     assert proxy.asdf_standard_requirement == SpecifierSet(">=1.4.0")
     assert proxy.converters == [ConverterProxy(c, proxy) for c in converters]
+    assert proxy.compressors == [MinimalCompressor]
     assert len(proxy.tags) == 1
     assert proxy.tags[0].tag_uri == "asdf://somewhere.org/extensions/full/tags/foo-1.0"
     assert proxy.types == []
@@ -209,6 +227,10 @@ def test_extension_proxy():
     # Should fail with a bad converter:
     with pytest.raises(TypeError):
         ExtensionProxy(FullExtension(converters=[object()]))
+
+    # Should fail with a bad compressor:
+    with pytest.raises(TypeError):
+        ExtensionProxy(FullExtension(compressors=[object()]))
 
     # Unparseable ASDF Standard requirement:
     with pytest.raises(ValueError):
@@ -555,6 +577,7 @@ extension_uri: asdf://somewhere.org/extensions/foo
         assert extension.legacy_class_names == []
         assert extension.asdf_standard_requirement is None
         assert extension.converters == []
+        assert extension.compressors == []
         assert extension.tags == []
 
         proxy = ExtensionProxy(extension)
@@ -562,6 +585,7 @@ extension_uri: asdf://somewhere.org/extensions/foo
         assert proxy.legacy_class_names == set()
         assert proxy.asdf_standard_requirement == SpecifierSet()
         assert proxy.converters == []
+        assert proxy.compressors == []
         assert proxy.tags == []
 
     with config_context() as config:
@@ -600,12 +624,14 @@ tags:
         extension = ManifestExtension.from_uri(
             "asdf://somewhere.org/extensions/foo",
             legacy_class_names=["foo.extension.LegacyExtension"],
-            converters=[converter]
+            converters=[converter],
+            compressors=[MinimalCompressor],
         )
         assert extension.extension_uri == "asdf://somewhere.org/extensions/foo"
         assert extension.legacy_class_names == ["foo.extension.LegacyExtension"]
         assert extension.asdf_standard_requirement == SpecifierSet(">=1.6.0,<2.0.0")
         assert extension.converters == [converter]
+        assert extension.compressors == [MinimalCompressor]
         assert len(extension.tags) == 2
         assert extension.tags[0] == "asdf://somewhere.org/tags/bar"
         assert extension.tags[1].tag_uri == "asdf://somewhere.org/tags/baz"
@@ -618,6 +644,7 @@ tags:
         assert proxy.legacy_class_names == {"foo.extension.LegacyExtension"}
         assert proxy.asdf_standard_requirement == SpecifierSet(">=1.6.0,<2.0.0")
         assert proxy.converters == [ConverterProxy(converter, proxy)]
+        assert proxy.compressors == [MinimalCompressor]
         assert len(proxy.tags) == 2
         assert proxy.tags[0].tag_uri == "asdf://somewhere.org/tags/bar"
         assert proxy.tags[1].tag_uri == "asdf://somewhere.org/tags/baz"